refactoring code:split collision logic from game panel, refactoring some rendering responsibilities
